requirements.txt
absl-py==1.4.0
accelerate==0.29.3
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==3.7.1
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
av==12.0.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpickle==3.0.0
clu==0.0.12
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.21
dashscope==1.17.0
decorator==5.1.1
distro==1.9.0
dm-reverb-nightly==0.9.0.dev20221205
dm-tree==0.1.8
einops==0.7.0
etils==1.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
flatbuffers==24.3.25
flax==0.8.2
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gast==0.5.4
gdown==5.2.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
gym==0.23.0
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
hydra-core==1.1.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
imgviz==1.7.5
immutabledict==4.2.0
importlib_resources==6.4.0
ipdb==0.13.13
ipython==8.23.0
jax==0.4.26
jaxlib==0.4.26
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
labelme==5.5.0
lazy_loader==0.4
libclang==18.1.1
lpips==0.1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
namex==0.0.7
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.2
onnxruntime==1.18.1
openai==1.2.3
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax @ git+https://github.com/google-deepmind/optax.git@478674f798343c93111b936ea3854398c3d35e74
optree==0.11.0
orbax-checkpoint==0.5.9
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
portpicker==1.6.0
progressbar2==4.4.2
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.6
pydantic==2.7.0
pydantic_core==2.18.1
pygame==2.1.0
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytorch-msssim==1.0.0
pytz==2024.1
PyYAML==6.0.1
QtPy==2.4.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rlds==0.1.7
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorflow==2.16.2
tensorflow-datasets==4.9.6
tensorflow-text==2.16.1
termcolor==2.4.0
tf-agents==0.14.0
tf-estimator-nightly==2.8.0.dev2021122109
tf_keras-nightly==2.16.0.dev2024021410
thop==0.1.1.post2209072238
tifffile==2024.6.18
tiktoken==0.6.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
traitlets==5.14.2
transformers==4.32.0
transformers-stream-generator==0.0.4
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.48
ultralytics-thop==2.0.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.14.1
yacs==0.1.8
yarl==1.9.4
zipp==3.18.1